code/Python/pyproject.toml
[tool.poetry]
name = "autogen-tests"
version = "0.1.0"
description = ""
authors = ["Thomas Tomow <toto_san@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
chromadb = "^0.5.3"
markdownify = "0.12.1"
autogen = "1.0.16"
agentops = "0.2.3"
alpha-vantage = "2.3.1"
arxiv = "2.1.0"
asttokens = "2.4.1"
autogenstudio = "0.0.56"
autopep8 = "2.3.1"
backports-tarfile = "1.2.0"
beautifulsoup4 = "4.12.3"
cachecontrol = "0.14.0"
cleo = "2.1.0"
crashtest = "0.4.1"
decorator = "5.1.1"
diskcache = "5.6.3"
distlib = "0.3.8"
distro = "1.9.0"
docker = "7.1.0"
docopt = "0.6.2"
dulwich = "0.21.7"
email-validator = "2.1.1"
executing = "2.0.1"
fastjsonschema = "2.20.0"
feedparser = "6.0.10"
filelock = "3.15.4"
flaml = "2.1.2"
frozendict = "2.4.4"
grpcio = "1.64.1"
html5lib = "1.1"
httptools = "0.6.1"
importlib-metadata = "7.1.0"
installer = "0.7.0"
ipython = "8.25.0"
jaraco-classes = "3.4.0"
jaraco-context = "5.3.0"
jaraco-functools = "4.0.1"
jedi = "0.19.1"
keyring = "25.2.1"
kubernetes = "30.1.0"
lxml = "5.2.2"
matplotlib = "3.9.0"
matplotlib-inline = "0.1.7"
monotonic = "1.6"
more-itertools = "10.3.0"
msgpack = "1.0.8"
multitasking = "0.0.11"
nh3 = "0.2.17"
openai = "1.34.0"
orjson = "3.10.5"
pandas = "2.2.2"
parso = "0.8.4"
peewee = "3.17.5"
pexpect = "4.9.0"
pkginfo = "1.11.1"
platformdirs = "4.2.2"
prompt-toolkit = "3.0.47"
psutil = "5.9.8"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyautogen = "^0.2.31"
pycodestyle = "2.12.0"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
rapidfuzz = "3.9.3"
readme-renderer = "43.0"
regex = "2024.5.15"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rfc3986 = "2.0.0"
setuptools = "70.1.0"
sgmllib3k = "1.0.0"
soupsieve = "2.5"
stack-data = "0.6.3"
termcolor = "2.4.0"
tiktoken = "0.7.0"
tomlkit = "0.12.5"
traitlets = "5.14.3"
trove-classifiers = "2024.5.22"
twine = "5.1.0"
typing = "3.7.4.3"
tzdata = "2024.1"
urllib3 = "2.2.2"
uvloop = "0.19.0"
virtualenv = "20.26.3"
watchfiles = "0.22.0"
wcwidth = "0.2.13"
webencodings = "0.5.1"
websockets = "12.0"
xattr = "1.1.0"
yfinance = "0.2.40"
pypdf = "^4.2.0"
sentence-transformers = "^3.0.1"
azure-search-documents = "^11.4.0"
azure-identity = "^1.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
